Get `ImportHighScoresCommand`'s components in a state they can be reused
